Autocomplete unknown room mention
