+++ src/app/components/editor/autocomplete/index.ts
@@ -1,2 +1,3 @@
 export * from './AutocompleteMenu';
 export * from './autocompleteQuery';
+export * from './RoomMentionAutocomplete';
