+++ src/app/utils/matrix.ts
@@ -1,0 +1,6 @@
+export const matchMxId = (id: string): RegExpMatchArray | null =>
+  id.match(/^([@!$+#])(\S+):(\S+)$/);
+
+export const validMxId = (id: string): boolean => !!matchMxId(id);
+
+export const getMxIdServer = (userId: string): string | undefined => matchMxId(userId)?.[3];
